homework9/services/order/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2
redis
retry
requests